Move importer test to gs1-importer and restore transpiler test
